Added snippets with lexer tweaks
Mostly just added the snippets that I like to use. On top of this, made
sure that tabbing is up to 4 spaces.

Did also a couple of additions to lexer, including commenting out the
single quote type string. Completely forgot of lifetime explicit
definitions when differncing with the C lexer.