semi-working ctags
So I got the tags sort of working. The one thing that is happening
now is that nothing is following their specific module and/or trait.
Probably has something to do with how the tags is referencing them.

Seeing as how I'm only using the std crate for now, I think I may build
a form that uses just the std crate for calls only. Also somehow get fn,
structs, and enums to only be able to be called using the operator `.`
and `::` for the traits, modules, and MAYBE crates.