An updated tags construction and api insertion
I decided today for whatever reason to continue on the work I set out
to do in building the `api` and `tags` files for adeptsense. I am
starting to get the hang of how Textadept fetches the autocomplete.

Right now, the api is not ready. I think instead of using the format in
Textadept, I will only use the prototypes for an api call. This is
because to make it read the documentation, I will need to make a full
fledge parser for each tag insertion. Not planning to do that any time
soon.

TL;DR API prototypes only going forward.