Major update to tags and api references.
With the incremental update of rust to v10, I decided to reference ALL
the crates with their respective tags and api file. Turns out it was
much easier than I thought it would be.

However, due to the heavy size of all the crates together, using
autocomplete with autosense on the tags does lag considerably. I may
end up to just referencing `tags_std` only but still maintaining the
`api` references for all the crates.