raw: refactoring get_project_name
A little needs to be said on how the fn is working now. Originally,
it was easier to just apply from `io.get_project_root()`, but as time
progress, Cargo projects are not synonymous with the root of vcs sw.
As such, it would be better advised to check the **pwd** for a
`Cargo.toml` file and then fail to doing the original
`io.get_project_root()` procedure.